refactor: plotly.py 생성
